remove explicit loops
Passing explicit loops to many asyncio apis is deprecated, and
it is discouraged in general. It was done initially because
`get_event_loop()` used to return the default event loop for
the thread, not the current event loop. This was a problem if
multiple event loops were running in a thread, as the default
loop may not be the loop you want. in 3.5.3+ this behavour was
changed, so `get_event_loop()` would return the current loop and
not the default loop, which is always what you want, so loops
don't need explicitly passed around anymore.

Resolves #122